requirements.txt
# Core dependencies
cryptography>=3.4.8
psutil>=5.8.0

# Browser extraction
pycryptodome>=3.15.0
keyring>=23.0.0

# GUI and desktop integration
dbus-python>=1.2.16
secretstorage>=3.3.0

# Optional: Keylogging and clipboard
pynput>=1.7.6
pyperclip>=1.8.2

# Optional: Memory analysis
volatility3>=2.4.0

# Reporting and output
jinja2>=3.0.0
markdown>=3.3.0

# Utilities
colorama>=0.4.4
tqdm>=4.62.0 